feat(plugin-sdk): split vulkan-jpeg — engine-free plugin/vulkan-jpeg + parked nvJPEG
Splits the dual-backend vulkan-jpeg along the cdylib-safety line established by the
audit:

- libs/vulkan-jpeg -> plugin/vulkan-jpeg (git mv, history preserved): the CPU JPEG
  parser + the native Vulkan-compute backend, now engine-free. Deps streamlib-plugin-sdk
  by real name (no streamlib facade); every RHI/color/error type resolves through
  streamlib_plugin_sdk::sdk::{rhi,context,color,error}. SimpleJpegDecoder builds the
  Vulkan-compute backend directly — the Auto/nvJPEG selection + the
  host_vulkan_device_arc caps probe (the unsound transit) are gone, as is the
  STREAMLIB_JPEG_BACKEND probe-skip workaround. cargo check green, zero streamlib-engine
  in the tree.
- nvJPEG -> libs/streamlib-engine/src/vulkan/_nvjpeg_impl_pending_/ (parked, NOT in the
  module tree, mirroring _apple_impl_pending_/): nvjpeg backend + resources + ffi + the
  host-RHI tests, plus a README pointing at #1206. nvJPEG legitimately belongs in the
  engine (raw HostVulkanDevice / OPAQUE_FD), has no blessed plugin API yet, and is
  re-integrated + cdylib-safe-exposed under #1206 (a perf follow-up, not
  competition-blocking). The misleading "host_vulkan_device_arc is cdylib-safe" comments
  are deleted.
- packages/jpeg migrated engine-free (streamlib facade -> streamlib-plugin-sdk, deps to
  =0.4.35-dev.9, streamlib::sdk::* -> streamlib_plugin_sdk::sdk::*, export_plugin!
  unchanged), dropping its STREAMLIB_JPEG_BACKEND plumbing.

Engine still compiles (nothing referenced the nvjpeg modules); xtask check-boundaries
clean. All five drone-racer pipeline packages are now off the engine.
