+++ Cargo.toml
@@ -3,6 +3,7 @@ resolver = "2"
 members = [
     "plugin/streamlib-error", # plugin/ zone: canonical Error + Result (engine-free)
     "plugin/streamlib-plugin-sdk", # plugin/ zone: engine-free authoring SDK (cdylib arm of the dual-mode views + glue)
+    "plugin/vulkan-jpeg", # plugin/ zone: engine-free JPEG decoder — CPU parser + Huffman entropy decode + fused Vulkan-compute kernel (cdylib-safe; nvJPEG backend parked in the engine)
     "xtask",                    # Development tasks (cargo xtask <command>)
     "libs/streamlib-sdk",       # Public authoring API surface — customer apps depend on this (Cargo name `streamlib`)
     "libs/streamlib-engine",    # Engine internals (RHI, IPC, surface-share, runtime executor) — private dep of the SDK
@@ -43,7 +44,6 @@ members = [
     # `[workspace]` self-root. Only `libs/*` + `xtask` form this workspace.
     # Build a package/example with `cd <dir> && cargo build`, or via
     # `streamlib pkg install` / runtime `Strategy::Registry` for packages.
-    "libs/vulkan-jpeg",                   # JPEG decoder: CPU parser + Huffman entropy decode + GPU compute kernel (compute kernel ships in a follow-up issue)
     # Exception to the "packages/* are not members" rule above: the
     # test-fixtures packages are INTERNAL test infrastructure, not
     # distributable units. The engine integration tests build them in-place
