+++ packages/jpeg/src/lib.rs
@@ -2,11 +2,11 @@
 // SPDX-License-Identifier: BUSL-1.1
 
 //! `@tatolab/jpeg` — GPU-backed JPEG decoder processor, thin wrapper
-//! around `libs/vulkan-jpeg::SimpleJpegDecoder`.
+//! around `vulkan-jpeg::SimpleJpegDecoder`.
 //!
 //! Linux-only today; the underlying primitive is Linux-only
-//! (VulkanComputeBackend + optional NvJpegBackend, both bound to
-//! `streamlib::sdk::context::GpuContextFullAccess`).
+//! (VulkanComputeBackend, bound to
+//! `streamlib_plugin_sdk::sdk::context::GpuContextFullAccess`).
 
 #[allow(non_snake_case, unused_imports, clippy::all)]
 pub mod _generated_ {
