+++ plugin/vulkan-jpeg/src/error.rs
