+++ plugin/vulkan-jpeg/src/lib.rs
@@ -26,8 +26,6 @@ pub mod backend;
 #[cfg(target_os = "linux")]
 pub mod kernel;
 #[cfg(target_os = "linux")]
-pub mod nvjpeg_backend;
-#[cfg(target_os = "linux")]
 pub mod simple_decoder;
 #[cfg(target_os = "linux")]
 pub mod vulkan_compute_backend;
@@ -50,11 +48,7 @@ pub use backend::{JpegBackendKind, JpegDecodeBackend};
 #[cfg(target_os = "linux")]
 pub use kernel::{JpegDecodeKernel, JPEG_DECODE_WORKGROUP_SIZE};
 #[cfg(target_os = "linux")]
-pub use nvjpeg_backend::NvJpegBackend;
-#[cfg(target_os = "linux")]
-pub use simple_decoder::{
-    JpegBackendPreference, JpegDecodeOutput, SimpleJpegDecoder, MAX_FRAMES_IN_FLIGHT,
-};
+pub use simple_decoder::{JpegDecodeOutput, SimpleJpegDecoder, MAX_FRAMES_IN_FLIGHT};
 #[cfg(target_os = "linux")]
 pub use vulkan_compute_backend::VulkanComputeBackend;
