+++ plugin/vulkan-jpeg/src/parser.rs
