+++ plugin/vulkan-jpeg/src/scan.rs
