+++ plugin/vulkan-jpeg/src/shaders/color_convert_common.glsl
