+++ plugin/vulkan-jpeg/src/shaders/jpeg_decode.comp
