+++ plugin/vulkan-jpeg/tests/integration.rs
